+++ marsworks/__init__.py
@@ -1,4 +1,4 @@
-__version__ = "0.3.1"
+__version__ = "0.3.3"
 from marsworks.client import *
 from marsworks.manifest import *
 from marsworks.origin import *